backend/python_ai_assistant/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
python-dotenv==1.0.1
openai==1.54.4
